Suggestion for a New Payment Details script
I refurbished the payment details script because according to the ER the payment details table doesn't need to reference the student ID since we'll use it for both staff and students. The payment id will be the primary key for the table and not unique as it will be referenced by the student and staff table.